Add support for passing memref arguments to MLIR.
This revision uses the python memref support introduced in https://reviews.llvm.org/D100077 to make a simple end-to-end example.

Therre is however a discrepancy between numpy which uses #bytes strides and memref which adopt the torch representation and use #elements strides. As a consequence some tests are temporarily disabled.

Bufferization is extended to properly work with a top-level function that does not have an MLIR caller.

This relies on LLVM @ ca0e250ec6e0f76e83358df5adc1ab061eed52f2

PiperOrigin-RevId: 367695329