[mlir][python[[sparse] Add exhaustive tests for sparse codegen.
This is the first CL in a serial for adding sparse codegen exhaustive tests.

Add input descriptor class and test descriptor class. The test descriptor class
has methods for getting the reference result as well as the test results for a
given combination of input values.

Define the annotation and codegen options used for the exhaustive tests.

Add basic tests for elementwise operations.

PiperOrigin-RevId: 385175902