[mlir][python[[sparse] Add dynamically generated test data.
Add two schemes to generate tensor data as test inputs: A dot scheme generates
non-zero values as scattered dots; a plane scheme generates non-zero values on
a plane.

Add a test name to the test descriptor. The test name is used to contruct the
directory name for the generated tensor input data.

Replace the static inputs for the tests with dynamic inputs. Make the test
input tensor dimensions more general.

PiperOrigin-RevId: 388307783