[mlir][python][sparse] Add the ability to run the reduction tests in parallel.
Add a command line option for specifying the number of processes used to run
the tests. When the number of processes is greater than 1, enable python
multiprocessing.

Change the routine that runs a test to return the name and status of the test
instead of printing out the test information. Add routines to run the tests
sequentially and parallelly.

PiperOrigin-RevId: 390371795