[mlir][python][sparse] Enhance the matrix multiplication tests.
Refactor the utilities to get the number of processes for running the tests and
invoke the tests. Move the utilities to the test common module.

Add the ability to run the matrix multiplication tests in parallel. Allow the
two input operands to have independent values for pointer bitwidth and index
bitwidth.

PiperOrigin-RevId: 391763607