[mlir][python][sparse] Enhance the sparse codegen exhaustive test tool.
Remove the need to preload the supporting library as the execution engine can
now take the library name as input and load the library.

Previously, the generated input tensors are stored in files and then read in
from the tests. Now, we use numpy arrays to buffer the generated input tensors,
thanks to the recently added convert operator in the sparse tensor dialect.

Fix a bug in the tensor generator which causes undesired density for the
generated tensor.

Make the elementwise op tests run in parallel. Decouple the bitwidth for the
operands.

PiperOrigin-RevId: 394687043