[mlir][python][sparse] Add more data types to the sparse codegen exhaustive test tool.
Previously, we only support data type float64. We now support more floating
point types as well as integer types.

Add TDType enums to represent the supported data types.

Add a data type parameter to each operation test, and add a routine to filter
out the unsupported data type combinations.

Add unit test for support_tensor_types.

Modify TestDesc to record each input tensor as a list of integers instead of a
numpy array. When we run a test for a given operation with a specific data
type, we construct the input tensor from the list of integers for the data type.
This allows the testing of the same operation with multiple data types using
the same TestDesc object.

PiperOrigin-RevId: 397085412