Retire codegen strategy on tensors that is subsumed by Driver.cpp
The codegen strategy on tensors has served its purpose and has been simplified into a Driver.cpp that follows MLIR core best practices re tablegen-based pass declaration.

Most tests have now become upstreamed and have become redundant with either upstream of with the python-bases test and search.

2 tests are mostly commented out until the tileanddistribute is integrated in the upstream strategy and/or sandbox driver.

test_constant.mlir is now fixed and does not depend on the strategy so it is reactivated for now.

PiperOrigin-RevId: 397979033