Implement einsum for contractions
Provide the mechanism to construct Linalg generic ops for arbitrary tensor
contractions using numpy-style einsum specification. This generalizes matmul
and matvec to arbitrary contractions with permutations. At this moment, we are
not generating all possible contractions as named ops since their number is too
large (thousands to millions depening on which kind of dimension permutations
are taken into account).

PiperOrigin-RevId: 411023387