Use collective arguments instead of arg tuples in ProblemDefinition
Refactor ProblemDefinition to use separate arguments:
  - sizes: Mapping[str, Any] that contains the mapping between names of the
    dimensions or other compile-time operation attributes (such as strides or
    dilations in convoluitons) and their numeric values;
  - types: Sequence[Type] for lists of types in the order they are accepted by
    the generated functions;
instead of a flat tuple of Python variadic arguments. This largely simplifies
the logic of all generic ProblemDefinitions that previously needed to split the
argument tuple based on type at an expense of slightly increasing (+2-3 lines)
the length of non-generic definitions. This tradeoff is acceptable since
non-generic definitions should be eventually generalized. It also removes the
unnecessary dependence on the order of specification of dimensions when running
the problem, enabling the possibility to factor out the test harness combining
data types, dimension sizes and experts, which will come next as a separate
commit.

Fix a number of issues in the process:
  - naming inconsistency in the reduction benchmark ('N' and 'K' were used
    interchangeably);
  - implement "TODO: FuncOp" in `build_op_under_context_manager`;
  - documentation in `mlir_types_builder` was copy-pasted from another function
    and thus wrong;
  - the `from collections.abc import Callable` is mostly useless for typing
    purposes as it cannot be subscripted, replace with `from typing import
    Callable`;
  - drop the unused `from typing import Type` that conflicts with
    `mlir.ir.Type`;
  - specify return type annotations in functions in ProblemDefinition
    subclasses.

PiperOrigin-RevId: 411575030