Stop relying on an ordered list of problem parameters in ProblemInstance
A previous commit rendered this constructor field irrelevant as the problem
parameters are accepted as a mapping anyway, the list was required only because
of internal data marshalling that has been refactored. Drop the key list
argument form the ProblemInstance constructor. When available, expose the list
of expected keys as a property on ProblemDefinition for convenience.