Make `TransformationList`s compose
This is the first step towards more composable `TransformationList` (aka
expert) API. It enables a construction of a `TransformationList` from a
list of Transformation subclasses without instantiating them as well as
a direct concatenation of two `TransformationList`s operating on the
same function/opname pair. Thanks to a metaclass, such
`TransformationList`s are concrete subclasses with a flat list of
variables amenable to the existing search mechanism. Each Transformation
now has its own list of variables that is forwarded down from the
`TransformationList`.

Further work is necessary to:

  - better compose TransformationList objects in addition to classes
    (these objects are not subject to search);
  - allow for direct concatenation of TransformationLists that operate o
    different function/opname pairs.