Make Transformation instances compose similarly to classes
A previous commit provided a fluent-style API for `Transformation` and
`TransformationList` subclasses to compose and create new
`TransformationList` subclasses parameterizable by a union of the
options of individual transformations. Implement the same for objects of
said classes. This makes the nicely chainable with the `then` function
that clearly indicates the order of transformations.

This makes the construction of custom transformation chains slightly
more verbose, although arguably more readable, by having each
transformation that takes parameters be listed instead of taking one
list of kwargs into a transformation list and dispatching them to
individual transformations. Further improvements are necessary in the
specification of transformation "variables", currently tied into
autotuning, to allow for a more concise specification of parameters.