Always use transforms for printing IR (#60)
`TransformationList`s used to support two ways of printing IR: using the
`Print` pseudo-transformation and using the various flags passed as
kwarg when creating the `TransformationList`. The latter is poorly
composable (e.g., transformations with different "print after all" flag
values have no natural composition). Drop it in favor of using the
`Print` transformation and the chainable `print_ir` function that
simplifies the insertion of `Print`s into the `TransformationList`.