Add utility to vector-predicate a tiled loop (#64)
This patch introduces a utility that applies vector predication to a tiled
loop. The tiled loop is assumed to be vectorized after predication using
the step of the tiled loop as vectorization factor. The transformation
is very simple for now: a new vector.predicate operation is added, which
encloses the loop body. The predicate ensures that the tiled loop will
be executed in a VF-way vector fashion without going beyond the upper bound
of the loop, even when the number of iterations is not multiple of VF.