Introduce transform-aware canonicalization
Introduce a new variant of the greedy pattern rewrite driver, adapted
from upstream MLIR since the class itself is an implementation detail
and cannot be derived. This variant is aware of the mapping between
values used in the Transform dialect and the operations and updates it
if the mapped operations are being updated by canonicalization patterns
(which does happen, e.g., after tiling to propagate static shape
information into types). This driver allows the transformation
interpreter to run a "canonicalization" pass without resorting to actual
passes and without risking to lose track of the operations that are
being transformed.