Run PDL matching patterns once per operation
Introduce a new helper function that wraps PatternApplicator so that
each pattern (in the case of Transform dialect, only the matching part
is actually relevant) is checked at most once against every operation.
No greedy recursive pattern application or worklist of newly created
operations is needed for this purpose. This allows us to stop relying
on the absence of the "matched" attribute in the IR to stop recursive
matching. The attribute itself is still added as a debugging aid.