Introduce transform-aware CSE (#125)
Introduce a new variant of common subexpression elimination, adapted
fomr upstream MLIR since the pass itself is an implementation detail and
cannot be derived. This variant can be run as a utility funciton and
does not plug into the pass infrastructure directly. It is also aware of
the operations being mapped to values used in the Linalg Transform
dialect and avoids deleting them so as not to break further
transformations.