Make run_tests preserve PYTHONPATH
run_tests.py sets the PYTHONPATH environment variable in a way that
ignores its current value, making it impossible for the caller to affect
it without modifying the script. Append the current value to the new
value instead.

Also dump error messages from the script so that the caller knows what
exactly went wrong.