Disallow more than one value use in the transform dialect
In the transform dialect, values correspond to an operation or a group
of operations in the IR being transformed. Applying a transformation
most likely changes the operation, so it becomes conceptually tricky to
relate the value to some operation. Instead, we can consider that the
value is "consumed" by the transformation operation, which may define
new values that map to the operations obtained as the result of
transformation for chaining purposes.