Avoid keeping pointers to deleted ops in transform interpreter (#229)
The transform interpreter maintains a mapping between the values
produced by the transform dialect operations and (sets of) operations in
the main IR to transform. Each transformation is likely to change the
tracked operation so the transform dialect diallows more than one use.
However, the pointer to the operation may still be kept in the mapping,
leading to use-after-free (or, in the case where the pointer address is
reused for a different operation, to the transform interpreter accessing
unexpected operations). Leverage the single-use restriction and clean up
the mapping immediately after the corresponding value is "consumed" by
the transform dialect operation.