Run enabler transformations in the transform dialect interpreter
These were initially omitted from the interpreter for simplicity. Add
them back to ensure parity with the code generation strategy. In the
latter, enablers seem optional, but in practice are run always except
for a single test verifying that they can be omitted.

This exposes the redundancy and the likely overhead in the enabler
transformations running between every other pairs of transformations,
while only some of the enablers are necessary for some of the
transformations (e.g. no need to simplify single-iteration loops if no
new loops are introduced, no need to hoist loop-invariant transfers if
no transfers are introduced, etc.) and that CSE effectively runs twice
in a row.