Check if the op is tracked in the listener
TrackingListener tries to find a new LinalgOp to use instead of the
currently tracked LinalgOp when the latter is being replaced. Check that
an op is actually tracked before looking for the replacement op. Beyond
this being potentially expensive, some untracked LinalgOps are replaced
with other (non-Linalg) ops, which can lead to assertion failures down
the road.