Iterators: Implement Map, ReduceByKey, Filter, and HashJoin. (#313)
This PR implements a handful more operators in the same spirit as the existing two.

This makes the set of operators expressive enough to implement about half of TPC-H (a standard database benchmark) except for the final sorting of the result. In addition to sorting (and TopK), the main missing operators are anti-join and outer join; further operators (such as left anti-join and left outer join) are also helpful for performance.