Introduce an extension mechanism for TransformState (#293)
This mechanism allows one to attach arbitrary typed extension objects to
the TransformState object, shared by ops implementing the
TransformOpInterface. Thus transform ops may communicate more
information than available in the state.

Use this mechanism instead of manually creating tracking listeners in
the Linalg Transform dialect interpreter. The current use is slightly
awkward as it requires to add and remove an extension in order to ensure
it reflects the most recent state of the payload IR, but it already
avoids reconstructing the internal structure of the tracking listener
for every canonicalization/CSE/enabler transformation that run
consecutively. Further work should remove the need to remove the
extension by notifying it about changes to the state.