Drop support for pass pipeline in the Tile transform and follow consequences.
Frpom now on, the Tile tranform can only be used with the transform dialect.
Ripping this bandaid has a number of implications and triggers a bunch of deletions that this PR implements.

A few bugs are exposed in the FIXME, they can be distributed and addressed as followups.

Followup cleanups can now easily occur on top of this PR, all the way to retiring most of the codegen strategy driver.