Update Github action for ccache to avoid cache flushing. (#498)
The version of the ccache Github action we are currently using uploads
the ccache even if the cache is empty. This can happen if a CI job is
cancelled at an unfortunate moment, such as this one:
https://github.com/google/iree-llvm-sandbox/runs/6447917749
(see that the cache size is 0.0 kB and that it is uploaded with the key
`ccache-Linux-buildtestasserts-2022-05-16T07:40:36.374Z`).
This means that subsequent builds start with an empty cache (as happened
here: https://github.com/google/iree-llvm-sandbox/runs/6447938528).

This issue in the Github action has been raised recently and fixed (see
https://github.com/hendrikmuhs/ccache-action/issues/45), which the
latest version includes.