SQL: Move schema from rel_opt into files (#512)
This PR moves the schema definition from rel_opt into the individual files. In this way, the files get more self-contained.

The only contextual things that are still part of rel_opt and need to be mapped in to the local context of the interpretation environment that generates ibis expressions from the files, are the numpy and ibis imports. I think the only way to get around this would be to actually parse things instead of interpreting them, which would mean we could also just forget about using ibis all together IMO.