SQL: add constantstreamOp to iterators mirror (#538)
This PR updates the iterators mirrored dialect with the constantstreamOp. This is mainly an operation used for testing purposes, which makes it a sensible first target to get to work.

Notice that we do not run any testcases for the mlir_conversion on our CI currently. With this PR there is no actual xdsl version that can handle this specific conversion. It has to be the xdsl checked out from here.