SQL: parametrize decimals (#551)
This PR adds the parametrization of decimals as requested in #546. Currently, there is no change in parameters when operations are performed. I decided to leave this out since, there isn't something like this done for other datatypes aswell. Consider int32. Adding two int32s should give an int33, but MLIR restricts results of addi to be the same as the input type (here).

I am still thinking about how to do this nicely, as I feel like I am in a weird mix of "the top-level should know" vs "aren't these implementation details?". I will probably push a version of something later to this PR.