Disable clang-format for long op class lists to keep formatting stable.
Changing the classes to the lists in `.Case<..>(..)` blocks may change
the indentation of the entire block due to how clang-format breaks
lines. This introduces unrelated changes for something that needs to be
changed frequently, so I think it is better to deactivate clang-format
here. Furthermore, this allows to have one class per line, which makes
the list and changes thereof easily readable.