Iterators: Improve analysis, type converter, and conversion driver (#554)
* Make signature and usage of computeStateType independent of Op type.

The current design of the IteratorAnalysis consists of a TypeSwitch
where each op (or at least group of op with the same number of operands)
has some glue code and then calls an op-specific function
`computeStateType`. This not only makes the signature of that function
dependent on the number of operands the corresponding op has, it also
distributes op-specific code over that function *and the TypeSwitch*,
both of which is unnecessary.

This commit changes the signature of `computeStateType` to use an array
of upstream state types, which may have zero, one, or more elements (one
for each operand).

* Small clean-up of IteratorsTypeConverter.

* Support non-iterator opperands in conversion to LLVM.

Until now, all iterator ops either have zero operands or operands that are
themselves iterators. However, I am about to introduce ops that have
"trraditional" (non-iterator) operands, which need to be handled
correctly.

* Disable clang-format for long op class lists to keep formatting stable.

Changing the classes to the lists in `.Case<..>(..)` blocks may change
the indentation of the entire block due to how clang-format breaks
lines. This introduces unrelated changes for something that needs to be
changed frequently, so I think it is better to deactivate clang-format
here. Furthermore, this allows to have one class per line, which makes
the list and changes thereof easily readable.

* Convert non-iterator operands with type converter.

A previous commit in this PR extends the custom conversion driver with
the ability to handle non-iterator operands by simply passing them
through unmodified. This commit converts them instead using the type
converter and unrealized conversion casts similar to what the normal
dialect conversion does.

* Use op adaptors in iterator-specific conversion logic.