SQL: add iterators scan op (#555)
This PR adds the iterators scan op, the way to actually lower FullTableScans to the iterators dialect. The PR relies on the version of the python bindings from #553. With this version, we can now go all the way through in terms of building the IR. Executing the code is still WIP.