SQL: small fixes for several parts of the pipeline (#556)
This PR upstreams several small fixes of issues I found on friday while making Q6 work (nearly) end-to-end. Notice that the handling for datetimes was adjusted to fit @ingomueller-net's way of generating the integers for the TPC-H benchmark.

Note that this PR depends on a UnitAttr, which I am currently upstreaming to xdsl. Therefore, it will currently break on the CI.