Delete IR on "low-level iterators" and related tests.
This IR was an early design attempt for lowering iterators to LLVM by
providing an intermediate level of abstraction. This was eventually
abandonded in favor of the current approach, which lowers to LLVM
functions and structs directly. While that worked, it is becoming more
and more clear that it bridges too high a gap, so we may want to
introduce such an intermediate level again at some point.