Decouple standalone project illustrating DB iterators from main project.
The template code we used at the very beginning of this project is now
not in use by the main project anymore. At some intermediate stage, we
used it as a "runtime" but that isn't the case anymore since we have a
proper lowering to LLVM. I think it may still be useful to keep this
code around as illustration. This commit moves it into its own folder
and decouples the compilation and testing workflow from the main
project.