SQL: add min, mean, max, count, and count_distinct reductions (#583)
This PR adds the min, mean, max, count, and count_distinct reductions. For the count, I decided to have an empty string as the col_name since we can take any column of the input operator to count, as all have the same number of rows. I also decided to rename mean to avg in rel_alg, as this corresponds to SQLs keyword (AFAIK).