SQL: fix projection pushdown and partial loading (#597)
This PR addresses an issue I introduced when "fixing" the handling of renames in the projection pushdown, in particular the projection simplifier. What I introduced was to restrictive to handle the TPC-H queries. So, I rewrote it to be able to handle some renames.

I also fixed a problem in the partial loading where projections would not be fused at all if there are complex projections present. I rewrote this to a version that does he partial loading, but also keeps the projection around.