SQL: add support for and (#599)
This PR adds support for the And node. In our current prototype, this can only occur in selections over lists of conditions, something like:

```python
table.filter([table.a == 5 & table.b == 7])
```

However, this is equivalent to writing:

```python
table.filter([table.a == 5, table.b == 7])
```

With this change, we lower to the latter form, if the former is encountered.