Implement utils for 1:N type conversion passes.
The upstream dialect conversion does not support 1:N type conversions.
This commit implements utils to build (poor-man's) dialect conversion
passes that do. The utils contist of a specialization of the
TypeConverter class that allows for N:1 target materializations, a
specialization of the RewritePattern class that automatically adds
appropriate unrealized casts supporting 1:N type conversions and
provides converted operands for implementing subclasses, and a
conversion driver that applies the provided patterns and replaces the
unrealized casts that haven't folded away with user-provided
materializations.