Only accept MemRefs with identity layout for tabular views.
We previously allowed those that match isStaticShapeAndContiguousRowMajor
but that function was removed from the public header in 46b90a7b5 of the
LLVM mono repo. The new restriction is more narrow than previously but
we only use memrefs that follow the restriction currently, and
implementing a test for the wider restriction is more complex/would
duplicate code.