Overhaul how types of referred to functions are verfied.
The previous implementation used ReferToOp, which 765ac6e9 in the LLVM
mono repo removed because it was not safe. The problem is that the op a
symbol refers to may be modified by a different thread during
verification, so the referred to op may not be inspected at that time.
The solution is to implement SymbolUserOpInterface, which allows for
deferred verification of anything related to the referred to symbol. The
downside is that all verification now happens in C++ instead of
tablegen.