[Iterators] Fix usage of updated build function of FuncOp.
This is (probably) due to dd74e6b6f in the LLVM mono repo, which adds
additional parameters to FuncOp, such that the build function that we
used before does not exist anymore. By not setting the visibility in the
construction, a different, still existing build function can be used.
The visibility can then be set by a setter, which (redundantly) was
already done anyways.