[Iterators] Update LLVM submodule to a version from January 20th (275ab24) and fix breakages. (#640)
* [LLVM] Update submodule to a version from September 27th (df7606a0).

* Use only prefixed accessors, since raw accessors are now disabled.

The switch was done in bb6d12b in the LLVM mono repo.

* Follow change of file and dialect names of arith dialect.

This was done in abc362a1 of the LLVM mono repo.

* Only accept MemRefs with identity layout for tabular views.

We previously allowed those that match isStaticShapeAndContiguousRowMajor
but that function was removed from the public header in 46b90a7b5 of the
LLVM mono repo. The new restriction is more narrow than previously but
we only use memrefs that follow the restriction currently, and
implementing a test for the wider restriction is more complex/would
duplicate code.

* Replace usages of deprecated is_splat with new all_equal.

This follows the change from 6fa87ec in the LLVM mono repo.

* [LLVM] Update submodule to a version from November 1st (cb84df21).

* Overhaul how types of referred to functions are verfied.

The previous implementation used ReferToOp, which 765ac6e9 in the LLVM
mono repo removed because it was not safe. The problem is that the op a
symbol refers to may be modified by a different thread during
verification, so the referred to op may not be inspected at that time.
The solution is to implement SymbolUserOpInterface, which allows for
deferred verification of anything related to the referred to symbol. The
downside is that all verification now happens in C++ instead of
tablegen.

* Replace removed NoSideEffect trait by new Pure trait.

The trait used until now was removed by 86771d0b in the LLVM mono repo.
If I understand things correctly, the Pure trait is somewhat more
restrictive but still applies for the affected ops.

* [LLVM] Update submodule to a version from November 23rd (61aed52c).

* Use new format to specify pipelines.

This becomes necessary with 66645a03f in the LLVM mono repo.

* [LLVM] Update submodule to a version from December 29th (1ec02146).

* [Iterators] Fix usage of updated build function of FuncOp.

This is (probably) due to dd74e6b6f in the LLVM mono repo, which adds
additional parameters to FuncOp, such that the build function that we
used before does not exist anymore. By not setting the visibility in the
construction, a different, still existing build function can be used.
The visibility can then be set by a setter, which (redundantly) was
already done anyways.

* [LLVM] Update submodule to a version from January 20th (275ab24).

* Use std::nullopt instead of llvm::None.

llvm::None has been deprecated by bd6a170 in the LLVM mono repo.

* Follow renaming of BlockAndValueMapping to IRMapping.

This was a change introduced in 4d67b278 of the LLVM mono repo.

* Use simpler constructor of SCF::IfOp.

This became necessary with 1125c5c of the LLVM mono repo, which removed
the constructors we used until now.