[Iterators] Replace UndefStateOp with CreateStateOp for iterator states. (#645)
The previous design was inspired from LLVM's structs and their way to
create an undef struct and then insert values. The downside of this
design is that it introduces the need to deal with undef states, which
gets complex if we want to allow arbitrary field types. By forcing the
existence of SSA values for all fields as arguments of a 'create' op, we
eliminate the possibility of undef states and push that issue up to the
user. Note that this means, in particular, that individual field values
can still be set to undef by creating a state with undef values if so
desired.