[Iterators] Replace utility builders for LLVM+SCF ops by upstreamed builders (NFC). (#653)
* Replace utility builders for LLVM ops by upstreamed builders (NFC).

We have implemented some time ago utility builder functions for LLVM's
ExtractValueOp and InsertValueOp with the plan to upstream them as
builders eventually. In the meantime, somebody else has done that, so we
can now use the upstreamed builders and remove our utility functions.

* Replace utility builders for SCF ops by upstreamed builders (NFC).

We have implemented some time ago utility builder functions for
SCF::WhileOp with the plan to upstream them as builders eventually. In
the meantime, somebody else has done that, so we can now use the
upstreamed builders and remove our utility functions.

* Remove our MLIRSupport library, which is now empty.