Introduce OneToNPatternRewriter for replacing ops.
The new rewriter class has an additional overlaod for replaceOp that
also inserts a cast back to the original types, which overcomes the
limit of the existing overlaps to require 1:1 (or no) type conversions.
This makes the interface of patterns nicer (they can use the rewriter as
usual instead of returning the desired replacements) and the conversion
back more explicit to the pattern author, which I guess improves clarity
as well.