[Iterators] Fix getOrCreateGlobalString for single-element Twines. (#665)
I had previously misunderstood (and hence misused) Twine::ToStringRef.
It fills the given SmallString *only if* the Twine isn't made from a
single element. I was using that SmallString blindly, which breaks in
the single-element case. Now, I differentiate between the SmallString as
storage and a StringRef that points to the value.