Simplify conversion and dialect library handling in CMake.
Turns out that dialect and converstion libraries are added to global
CMake variables if they are created with the right `add_mlir_*_library`
functions. This commit uses those functions were previously a more
generic function was used, and removes the explicit listing of dialect
and conversion libraries, which becomes unnecessary thanks to the global
lists.