[Triton][LLVM] Update submodules to latest versions as of June 6. (#746)
* Update LLVM (744e589c) and Triton (a5fceea1) submodules.
* Build Triton with self-built LLVM in CI and README.

This allows to update LLVM more easily: since Triton depends on LLVM and
we include part of it, our and Triton's LLVM dependency need to be
updated more or less in lock step. Until now, we used the official LLVM
dependency of Triton, which has a rather lengthy release process. By
using the LLVM dependency that we build ourselves, we control the whole
process and can update more quickly. (We do need to update the llvm-head
branch of Triton, though.)