Make calling kernels from other kernels work.
It is possible in Triton to call a jitted function from other jitted
function. In this case, the called function seems to be pulled into the
MLIR module of the calling function. To do that, the AST visitor does an
isinstance(JITFunction) test. However, before this commit, that test
failed because the JITFunction we use in our copied code is a different
Python class than the one that is tested for with the same name. This
commit thus changes the copied code such that that class inherits from
the original one (and then overwrites all of its content).