Activate `-eo pipefail` for benchmarks to make sure failures are raised.
One command in the CI step that test the benchmarks uses a pipe. Without
the `-eo pipefail` options, the exit code of a pipe statement is that of
the last of the commands, which means that failures in the previous
commands pass unnoticed. This actually happened for that CI step, which
has been failing for a long time.