[CI] Make sure failing benchmark test fail CI and fix current issue. (#757)
* Activate `-eo pipefail` for benchmarks to make sure failures are raised.

One command in the CI step that test the benchmarks uses a pipe. Without
the `-eo pipefail` options, the exit code of a pipe statement is that of
the last of the commands, which means that failures in the previous
commands pass unnoticed. This actually happened for that CI step, which
has been failing for a long time for various reasons:

* Fix various issues in the benchmark script that had passed unnoticed
  over time.
* Set environment variables with paths to runner util libraries.

Some unrelated minor fixes:

* Remove unused imports.
* Apply yapf formatting.