[CMake] Add missing dependencies to Triton tablegen targets. (#762)
Previously, we were missing a dependency from various uses of the
Triton-related tablegen files to the targets that produce them. To add
them, it is sufficient to change the order of the conversion tablegen
targets of this project. Curiously, the way to define these targets with
add_public_tablegen_target implicitly creates a dependencie of the new
target to the previous targets defined previously (in that file?). We
thus define the target MLIRStructuredConversionIncGen last such that all
targets already depending on it now also depend on the Triton-related
ones.