[Triton] Use ConvertOpToLLVMPattern where appropriate.
In ConvertTritonToLLVM, some patterns need to deal with types that get
converted to LLVM. This PR turns these patterns into
ConvertOpToLLVMPatterns, which makes the LLVMTypeConverter available
there (instead of the TypeConverter base). While this isn't used in the
implementations yet, #767 will require it, plus, I believe that this
distinction makes sense even conceptually.