[Triton] Overhaul and extend how pointer types are handled. (#767)
This concerns the conversion of `tt.addptr`, `tt.load`, and `tt.store`.
Previously, the implementation of conversion patterns assumed that the
`tt.ptr` operands always pointed to integer or float types, or even just
integer types. This commit makes the handling more consistent in that
the type conversion only works for `tt.ptr` where the pointee type can
be converted to an LLVM-compatible type, uses the converted pointee type
everywhere where appropriate (instead of assuming that no conversion is
necessary), and now also handles pointers of pointers correctly.