[LLVM][Triton] Update to latest versions as of July 13. (NFC) (#788)
* Update to latest version of LLVM (701c4adc) and Triton (4508d87c).

Importantly, this gets rid of the dependency of my fork of Triton since
the commit we depended on has now landed in the llvm-head branch.

* Compile Triton with GCC since Clang fails with warning and -Werror.

This isn't ideal since we can't link with LLD anymore, so the CI steps
takes longer. Hopefully the warning will be fixed and we can switch back.

* [Triton] Adapt call to ast_to_ttir to updstream change.

6d1285e1@openai/triton introduced the new argument `arch` to that
function. I have not fully understood how this argument works; it looks
like different architectures are differentiated by the type of the that
argument. Simply passing `None` seems to work.