[Substrait] Extend FieldReferenceOp to the `expression` case.
Until now, we only had the case of the `FieldReference` message
implemented that referend to a whole row of the current input relation
(i.e., the `root_reference` case). This PR extends the corresponding op
to support also the `expression` case, where an arbitrary expression
(producing a container) can be the input.