[Substrait] Use `substrait-cpp` instead of `substrait`. (#808)
The latter project only consists of the protobuf files, which required
us to create our own CMake logic to include those. The latter not only
provides the generated protobuf files as a ready-to-use library target
but also many more utilities for dealing with plans (including a
human-readable text format), which we'll probably want to use sooner or
later.