Rewrite export of `literal` ops.
This was raised by @jpienaar in the review of #818. One comment was
about the semantics of the numbers. This made me realize that
`set_boolean` is the wrong way for creating a protobuf message with an
`i32`. This passed unnoticed because we can't express plans with 32-bit
literals yet. I added a TODO for adding such as test. I also fixed the
export to the best of my knowledge. In passing, I also addressed the
other comment: the code now only tests properties of existing types
instead of creating types and then testing for equality, which should be
cheaper.

Signed-off-by: Ingo Müller <ingomueller@google.com>