Refactor how `emit` is imported and exported.
The first attempt had been rather akward, requiring the import and
export of each specific op to cooperate with the corresponding logic of
the `emit` op. This was due to the fact that there is no out-of-the-box
way to access the `RelCommon` message of the `rel_type` message without
being specific to the `rel_type` case. In this version, two utility
functions provide access to that message (one read-only, one mutable).
This allows the import and export of `emit` to be local, which removes a
whole class of difficult to detect bugs.

Signed-off-by: Ingo Müller <ingomueller@google.com>