[Substrait] Implement `CallOp` for `ScalarFunction` message.
This allows to use the externally declared functions from #853 as scalar
function calls (which have the semantics of a typical function call).
The current design anticipates that aggregate and window functions can
be modelled with the same op, but future PR will need to show if and how
that is possible.

Signed-off-by: Ingo Müller <ingomueller@google.com>