Fix tests of emit deduplication pass.
In the meantime, `main` contains tests that used `func.call` ops to
prevent further pattern application. This breaks with this PR, which
makes `plan`s a `SymbolTable`, which become the closest symbol table of
the `func.call`s, such that those symbols cannot be found anymore. As a
solution, this commit changes the unit tests to use an opaque op from an
unregistered dialect.

Signed-off-by: Ingo Müller <ingomueller@google.com>